Consumer-Retail-AI-Tech-Sparks--main/Consumer retail/reviewiq/requirements.txt
# After pip install -r requirements.txt, also run:
# python -m spacy download en_core_web_sm
mistralai>=1.0.0
streamlit>=1.35.0
fastapi>=0.111.0
uvicorn>=0.30.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
scikit-learn>=1.5.0
plotly>=5.22.0
ftfy>=6.2.0
langdetect>=1.0.9
emoji>=2.12.0
spacy>=3.7.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
requests>=2.32.0
networkx>=3.0.0
pyvis>=0.3.0
sqlalchemy>=2.0.0
deep_translator>=1.11.4
